+++ packages/navbars/src/components/Tabs/__tests__/Tabs.test.tsx
@@ -4,7 +4,8 @@ import React from 'react';
 import Tabs from '../Tabs';
 
 describe('Tabs', () => {
-  const list = [{value: '1', onPress: () => null}];
+  const list = [{label: '1', value: 1}];
+  const onPress = () => null;
 
   test('render correctly', () => {
     const {toJSON} = render(<Tabs list={list} activeValue={'1'} />);
@@ -15,10 +16,13 @@ describe('Tabs', () => {
       <Tabs
         list={list}
         activeValue={'1'}
+        onPressCommon={onPress}
         backgroundColorTab={'#000'}
+        backgroundColorActiveTab={'#ff0000'}
         contentContainerStyle={{flex: 1}}
         containerStyle={{flex: 2}}
         fontTab={'Regular-Error-H3'}
+        fontActiveTab={'Regular-White-H3'}
       />,
     );
     expect(toJSON()).toMatchSnapshot();