+++ packages/navbars/src/types.ts
@@ -3,23 +3,29 @@ import {ViewStyle} from 'react-native';
 import {TypographyProp} from '@npm/mobydick-typography';
 
 export interface ITab {
-  value: string;
+  value: string | number;
+  label: string;
   onPress?(): void;
   leftIcon?: ReactElement;
   rightIcon?: ReactElement;
 }
 
-export interface ITabProps {
+interface ITabCommon {
+  fontTab?: TypographyProp;
+  fontActiveTab?: TypographyProp;
+  backgroundColorTab?: string;
+  backgroundColorActiveTab?: string;
+  onPressCommon?(item: ITab): void;
+}
+
+export interface ITabProps extends ITabCommon {
   item: ITab;
   active: boolean;
-  backgroundColorTab?: string;
-  fontTab?: TypographyProp;
 }
-export interface ITabsProps {
+
+export interface ITabsProps extends ITabCommon {
   list: ITab[];
-  activeValue: string;
-  backgroundColorTab?: string;
-  fontTab?: TypographyProp;
+  activeValue: string | number;
   containerStyle?: ViewStyle | ViewStyle[];
   contentContainerStyle?: ViewStyle | ViewStyle[];
 }