+++ packages/popups/src/components/ActionSheetBase/ActionSheetBase.tsx
@@ -22,7 +22,9 @@ const ActionSheetBase: FC<
   const [styles] = useStyles(stylesCreate);
 
   return (
-    <PopupBase onClose={onClose} overlayStyle={overlayStyle}>
+    <PopupBase
+      onClose={onClose}
+      overlayStyle={[styles.overlayStyle, overlayStyle]}>
       <View
         style={[styles.containerStyle, containerStyle]}
         onStartShouldSetResponder={returnTrue}>