+++ storybook/stories/popups/Popup/ExampleActionSheet.tsx
@@ -18,7 +18,6 @@ const applyText = 'Применить';
 const deniedText = 'Отказано';
 const coordinationText = 'Согласование';
 const agreedText = 'Согласован';
-const textFontAccentM = 'Medium-Accent-M';
 const textFontErrorL = 'Regular-Error-L';
 
 const leftIconAccent = (name: SimpleIconName) => {
@@ -30,6 +29,8 @@ const ExampleActionSheet: FC<IContentProps> = props => {
   const [checkboxList, setCheckboxList] = useState<string[]>([]);
   const isShowLeftIcon = boolean('show left icon', true);
   const leftIconName = select('name left icon', iconNames, 'icon-settings');
+  const disabled = boolean('disabled', false);
+  const leftIcon = isShowLeftIcon ? leftIconAccent(leftIconName) : undefined;
 
   const onPressCheckbox = (title: string) => {
     if (checkboxList.find(item => item === title)) {
@@ -66,7 +67,7 @@ const ExampleActionSheet: FC<IContentProps> = props => {
           <ActionSheetBase.Item
             itemType={IItemType.cancelItem}
             title={cancelText}
-            textFont={textFontAccentM}
+            disabled={disabled}
             onPress={onClose}
           />
         </ActionSheetBase>
@@ -81,7 +82,7 @@ const ExampleActionSheet: FC<IContentProps> = props => {
               onPressRadio(deniedText);
             }}
             radio={radio}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.innerItem}
@@ -90,7 +91,7 @@ const ExampleActionSheet: FC<IContentProps> = props => {
               onPressRadio(coordinationText);
             }}
             radio={radio}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.lastItem}
@@ -99,11 +100,11 @@ const ExampleActionSheet: FC<IContentProps> = props => {
               onPressRadio(agreedText);
             }}
             radio={radio}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.cancelItem}
-            textFont={textFontAccentM}
+            disabled={disabled}
             title={applyText}
             onPress={onClose}
           />
@@ -119,7 +120,7 @@ const ExampleActionSheet: FC<IContentProps> = props => {
               onPressCheckbox(deniedText);
             }}
             checkboxList={checkboxList}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.innerItem}
@@ -128,7 +129,7 @@ const ExampleActionSheet: FC<IContentProps> = props => {
               onPressCheckbox(coordinationText);
             }}
             checkboxList={checkboxList}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.lastItem}
@@ -137,11 +138,11 @@ const ExampleActionSheet: FC<IContentProps> = props => {
               onPressCheckbox(agreedText);
             }}
             checkboxList={checkboxList}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.cancelItem}
-            textFont={textFontAccentM}
+            disabled={disabled}
             title={applyText}
             onPress={onClose}
           />
@@ -158,8 +159,8 @@ const ExampleActionSheet: FC<IContentProps> = props => {
           />
           <ActionSheetBase.Item
             itemType={IItemType.cancelItem}
-            textFont={textFontAccentM}
             title={cancelText}
+            disabled={disabled}
             onPress={onClose}
           />
         </ActionSheetBase>
@@ -171,33 +172,33 @@ const ExampleActionSheet: FC<IContentProps> = props => {
           <ActionSheetBase.Item
             itemType={IItemType.firstItem}
             title={settingsText}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.innerItem}
             title={settingsText}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.innerItem}
             title={settingsText}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.innerItem}
             title={settingsText}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.lastItem}
             title={settingsText}
             onPress={onClose}
-            leftIcon={isShowLeftIcon ? leftIconAccent(leftIconName) : undefined}
+            leftIcon={leftIcon}
           />
           <ActionSheetBase.Item
             itemType={IItemType.cancelItem}
-            textFont={textFontAccentM}
             title={resetText}
+            disabled={disabled}
             onPress={onClose}
           />
         </ActionSheetBase>